refactor(sdk): export argument types from lowlevel
