+++ packages/new-polymath-sdk/src/LowLevel/types.ts
@@ -41,8 +41,16 @@ export interface Erc20DividendDepositedEvent {
 }
 
 export enum DividendModuleTypes {
-  Erc20,
-  Eth,
+  Erc20 = 'erc20',
+  Eth = 'eth',
+}
+
+export enum ModuleTypes {
+  Permission = 1,
+  Transfer,
+  Sto,
+  Dividends,
+  Burn,
 }
 
 export interface Dividend {
@@ -71,4 +79,98 @@ export interface Checkpoint {
   investorBalances: InvestorBalance[];
   totalSupply: BigNumber;
   createdAt: Date;
+}
+
+// Transaction argument types
+
+export interface SetWithholdingArgs {
+  investors: string[];
+  percentages: number[];
+}
+
+export interface ReclaimDividendArgs {
+  dividendIndex: number;
+}
+
+export interface WithdrawWithholdingArgs {
+  dividendIndex: number;
+}
+
+export interface CreateErc20DividendArgs {
+  maturityDate: Date;
+  expiryDate: Date;
+  tokenAddress: string;
+  amount: BigNumber;
+  checkpointId: number;
+  name: string;
+  excludedAddresses?: string[];
+}
+
+export interface CreateEtherDividendArgs {
+  maturityDate: Date;
+  expiryDate: Date;
+  amount: BigNumber;
+  checkpointId: number;
+  name: string;
+  excludedAddresses?: string[];
+}
+
+export interface GetModuleFactoryAddressArgs {
+  moduleName: string;
+  moduleType: ModuleTypes;
+  tokenAddress: string;
+}
+
+export interface GetModulesByTypeAndTokenArgs {
+  moduleType: ModuleTypes;
+  tokenAddress: string;
+}
+
+export interface GetAddressArgs {
+  contractName: string;
+}
+
+export interface AllowanceArgs {
+  tokenOwner: string;
+  spender: string;
+}
+
+export interface GetTokensArgs {
+  amount: BigNumber;
+  recipient: string;
+}
+
+export interface BalanceOfArgs {
+  address: string;
+}
+
+export interface ApproveArgs {
+  spender: string;
+  amount: BigNumber;
+}
+
+export interface AddDividendsModuleArgs {
+  type: DividendModuleTypes;
+  wallet: string;
+}
+
+export interface GetModuleAddressArgs {
+  name: string;
+}
+
+export interface RegisterTickerArgs {
+  owner: string;
+  ticker: string;
+  tokenName: string;
+}
+
+export interface GenerateSecurityTokenArgs {
+  tokenName: string;
+  ticker: string;
+  tokenDetails: string;
+  divisible: boolean;
+}
+
+export interface GetSecurityTokenArgs {
+  ticker: string;
 }
